tests/wpt/web-platform-tests/css/css-text/parsing/text-justify-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-justify with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-justify">
<meta name="assert" content="text-justify supports the full grammar 'auto | none | inter-word | inter-character'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-justify", "auto");
test_valid_value("text-justify", "none");
test_valid_value("text-justify", "inter-word");
test_valid_value("text-justify", "inter-character");
</script>
</body>
</html>
